Fixed issue with parallel termination.

+++ source/System/Threading/Tasks/Parallel.js
@@ -104,12 +104,14 @@
         function getNew(key, url) {
             var worker = new Worker_1.default(url);
             worker.__key = key;
-            worker.dispose = function () {
-                worker.terminate();
-                worker.onmessage = null;
-                worker.onerror = null;
-                worker.dispose = null;
-            };
+            if (!worker.dispose) {
+                worker.dispose = function () {
+                    worker.onmessage = null;
+                    worker.onerror = null;
+                    worker.dispose = null;
+                    worker.terminate();
+                };
+            }
             return worker;
         }
         workers.getNew = getNew;
