+++ source/System/Threading/Tasks/Parallel.ts
@@ -153,13 +153,16 @@ module workers
 	{
 		const worker:any = new Worker(url);
 		worker.__key = key;
-		worker.dispose = () =>
+		if(!worker.dispose)
 		{
-			worker.terminate();
-			worker.onmessage = null;
-			worker.onerror = null;
-			worker.dispose = null;
-		};
+			worker.dispose = () =>
+			{
+				worker.onmessage = null;
+				worker.onerror = null;
+				worker.dispose = null;
+				worker.terminate();
+			};
+		}
 		return worker;
 	}
 }
